Int64 -> Int

+++ test/nullablevector.jl
@@ -224,25 +224,25 @@ module TestNullableVector
     empty!(X)
     @test isempty(X)
 
-    @test typeof(vcat(NullableArray(1:2), 3:4)) == NullableArray{Int64,1}
-    @test typeof(vcat(1:2, NullableArray(3:4))) == NullableArray{Int64,1}
-    @test typeof(vcat(NullableArray([1 2]), [3 4])) == NullableArray{Int64,2}
-    @test typeof(vcat([1 2], NullableArray([3 4]))) == NullableArray{Int64,2}
-    @test typeof(vcat(NullableArray(1:2), 3:4, 5:6)) == NullableArray{Int64,1}
-    @test typeof(vcat(1:2, NullableArray(3:4), 5:6)) == NullableArray{Int64,1}
-    @test typeof(vcat(1:2, 3:4, NullableArray(5:6))) == NullableArray{Int64,1}
-    @test typeof(vcat(NullableArray([1 2]), [3 4], [5 6])) == NullableArray{Int64,2}
-    @test typeof(vcat([1 2], NullableArray([3 4]), [5 6])) == NullableArray{Int64,2}
-    @test typeof(vcat([1 2], [3 4], NullableArray([5 6]))) == NullableArray{Int64,2}
+    @test typeof(vcat(NullableArray(1:2), 3:4)) == NullableArray{Int,1}
+    @test typeof(vcat(1:2, NullableArray(3:4))) == NullableArray{Int,1}
+    @test typeof(vcat(NullableArray([1 2]), [3 4])) == NullableArray{Int,2}
+    @test typeof(vcat([1 2], NullableArray([3 4]))) == NullableArray{Int,2}
+    @test typeof(vcat(NullableArray(1:2), 3:4, 5:6)) == NullableArray{Int,1}
+    @test typeof(vcat(1:2, NullableArray(3:4), 5:6)) == NullableArray{Int,1}
+    @test typeof(vcat(1:2, 3:4, NullableArray(5:6))) == NullableArray{Int,1}
+    @test typeof(vcat(NullableArray([1 2]), [3 4], [5 6])) == NullableArray{Int,2}
+    @test typeof(vcat([1 2], NullableArray([3 4]), [5 6])) == NullableArray{Int,2}
+    @test typeof(vcat([1 2], [3 4], NullableArray([5 6]))) == NullableArray{Int,2}
 
-    @test typeof(hcat(NullableArray(1:2), 3:4)) == NullableArray{Int64,2}
-    @test typeof(hcat(1:2, NullableArray(3:4))) == NullableArray{Int64,2}
-    @test typeof(hcat(NullableArray([1 2]), [3 4])) == NullableArray{Int64,2}
-    @test typeof(hcat([1 2], NullableArray([3 4]))) == NullableArray{Int64,2}
-    @test typeof(hcat(NullableArray(1:2), 3:4, 5:6)) == NullableArray{Int64,2}
-    @test typeof(hcat(1:2, NullableArray(3:4), 5:6)) == NullableArray{Int64,2}
-    @test typeof(hcat(1:2, 3:4, NullableArray(5:6))) == NullableArray{Int64,2}
-    @test typeof(hcat(NullableArray([1 2]), [3 4], [5 6])) == NullableArray{Int64,2}
-    @test typeof(hcat([1 2], NullableArray([3 4]), [5 6])) == NullableArray{Int64,2}
-    @test typeof(hcat([1 2], [3 4], NullableArray([5 6]))) == NullableArray{Int64,2}
+    @test typeof(hcat(NullableArray(1:2), 3:4)) == NullableArray{Int,2}
+    @test typeof(hcat(1:2, NullableArray(3:4))) == NullableArray{Int,2}
+    @test typeof(hcat(NullableArray([1 2]), [3 4])) == NullableArray{Int,2}
+    @test typeof(hcat([1 2], NullableArray([3 4]))) == NullableArray{Int,2}
+    @test typeof(hcat(NullableArray(1:2), 3:4, 5:6)) == NullableArray{Int,2}
+    @test typeof(hcat(1:2, NullableArray(3:4), 5:6)) == NullableArray{Int,2}
+    @test typeof(hcat(1:2, 3:4, NullableArray(5:6))) == NullableArray{Int,2}
+    @test typeof(hcat(NullableArray([1 2]), [3 4], [5 6])) == NullableArray{Int,2}
+    @test typeof(hcat([1 2], NullableArray([3 4]), [5 6])) == NullableArray{Int,2}
+    @test typeof(hcat([1 2], [3 4], NullableArray([5 6]))) == NullableArray{Int,2}
 end
